Make 2D geometry to use Double again, while keeping Float geometry for 3D (#1868)

+++ korge-core/src/commonMain/kotlin/korlibs/image/font/GlyphPath.kt
@@ -12,7 +12,7 @@ data class GlyphPath(
     var bitmapOffset: Point = Point(0, 0),
     var bitmapScale: Scale = Scale(1, 1),
     var transform: Matrix = Matrix(),
-    var scale: Float = 1f
+    var scale: Double = 1.0
 ) : Drawable {
     val isOnlyPath get() = bitmap == null && colorShape == null
